capture: allow to join several mcast groups
get_sdp.py write a conf file used by the capture script which contains
the info of the 1st video, audio and ancillary provided the Embrionix
source.

capture.sh still can process manual arguments, for instance if the
source is not an Embrionix sender and no conf file can be created.

socat is replaced smcroute for easier joining.