transcoder: restart when pass file is getting too large
libx264 ouputs a lot of stats (~10GB in a day). This pass file is
useless in a live application but it turns out there is no way to avoid
this file from being written and it can't be redirected to /dev/null.
ffmpeg is prone to fail the file is too large:
        error: ratecontrol_init: can't open stats file

'monitor' cmd now checks the file size and restart the transcoder if
needed.